firetec-multithread/requirements.txt
# Dependências do projeto FireTec Multithread
# Python 3.8+

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Geolocalização
geopy>=2.3.0
shapely>=2.0.0
polycircles>=0.3.0

# Mapas e KML
simplekml>=1.3.6

# Audio
gTTS>=2.3.0
pydub>=0.25.1
imageio-ffmpeg>=0.4.9

# CAP (Common Alerting Protocol)
capparser>=0.1.0

# Logging
colorlog>=6.7.0

# Configuração
python-dotenv>=1.0.0
pyyaml>=6.0

# Testes (opcional)
pytest>=7.4.0
pytest-cov>=4.1.0
